+++ src/main/kotlin/com/partymakers/shareparty/party/domain/usecase/CreatePartyRoomUseCase.kt
@@ -1,5 +1,5 @@
 package com.partymakers.shareparty.party.domain.usecase
 
 interface CreatePartyRoomUseCase {
-    operator fun invoke(partyName: String): Int
+    operator fun invoke(partyName: String): Long
 }
