+++ src/main/kotlin/com/partymakers/shareparty/party/domain/usecase/CreatePartyRoomUseCaseImpl.kt
@@ -8,6 +8,6 @@ internal class CreatePartyRoomUseCaseImpl(
     private val repository: PartyRoomRepository
 ) : CreatePartyRoomUseCase {
 
-    override fun invoke(partyName: String): Int =
+    override fun invoke(partyName: String): Long =
         repository.create(partyName)
 }
